Reorganized models into separate files
