+++ Sakurajima/models/__init__.py
@@ -1,0 +1,9 @@
+from Sakurajima.models.base_models import *
+from Sakurajima.models.chronicle import *
+from Sakurajima.models.media import *
+from Sakurajima.models.notification import *
+from Sakurajima.models.recommendation import *
+from Sakurajima.models.relation import *
+from Sakurajima.models.stats import *
+from Sakurajima.models.user_models import *
+from Sakurajima.models.watchlist import *
